CI speedup: run unit tests in series, and ensure that scripts arent repeated for different figs

+++ tests/test.m
@@ -4,10 +4,10 @@ classdef (SharedTestFixtures={ ...
     % class based unit tests, as in https://www.mathworks.com/help/matlab/matlab_prog/class-based-unit-tests.html
     
     properties (Constant)
-        run_slow_tests = false;
+        run_slow_tests = true;
 
-        slow_figs = feval(@(all_figs) all_figs(~ismember(all_figs,[16:21, 24:25])), 1:36)
-        slow_tabs = feval(@(all_tabs) all_tabs(~ismember(all_tabs, 7:8          )), 1:8)
+        slow_figs = [16:21, 24:25];
+        slow_tabs = 7:8;
     end
 
     properties
@@ -23,7 +23,12 @@ classdef (SharedTestFixtures={ ...
         actual_wecsim
         econ_fig_wecsim
 
-        uuid   
+        uuid
+
+        fig_success
+        tab_success
+        fig_output
+        tab_output
     end
 
     % inputs for tests, including passing tolerances
@@ -39,7 +44,7 @@ classdef (SharedTestFixtures={ ...
     % helper methods to enumerate all figures and tables
     methods (Static)
         function which_fig_struct = enumerateFigs()
-            [~,~,~,num_figs,num_tabs,fig_names,~] = all_figures( [],[] );
+            [~,~,~,~,num_figs,num_tabs,fig_names,~] = all_figures( [],[] );
 
             which_figs_vec = [1:num_figs zeros(1, num_tabs)];
             none = strcat(repmat({'none'},1,num_tabs), string(1:num_tabs));
@@ -49,7 +54,7 @@ classdef (SharedTestFixtures={ ...
             which_fig_struct = cell2struct(which_figs_cell,fig_names,2);
         end
         function which_tab_struct = enumerateTabs()
-            [~,~,~,num_figs,num_tabs,~,tab_names] = all_figures( [],[] );
+            [~,~,~,~,num_figs,num_tabs,~,tab_names] = all_figures( [],[] );
 
             which_tabs_vec = [zeros(1,num_figs), 1:num_tabs];
             none = strcat(repmat({'none'},1,num_figs), string(1:num_figs));
@@ -87,6 +92,40 @@ classdef (SharedTestFixtures={ ...
             % required to prevent file overalps for generated code
             testCase.uuid = parallel.pool.Constant(@() char(matlab.lang.internal.uuid()));
         end
+
+        function runAllFigsTabs(testCase)
+            [~,~,~,~,num_figs,num_tabs] = all_figures( [], [] );
+
+            all_figs = 1:num_figs;
+            all_tabs = 1:num_tabs;
+            if ~testCase.run_slow_tests
+                run_figs = all_figs(~ismember(all_figs,testCase.slow_figs));
+                run_tabs = all_tabs(~ismember(all_tabs,testCase.slow_tabs));
+            else
+                run_figs = all_figs;
+                run_tabs = all_tabs;
+            end
+
+            [f_success_run,t_success_run,...
+             f_output_run, t_output_run] = all_figures( run_figs, run_tabs, testCase.uuid );
+
+            f_success = cell(1,num_figs);
+            f_success(run_figs) = f_success_run;
+
+            t_success = cell(1,num_tabs);
+            t_success(run_tabs) = t_success_run;
+
+            f_output = gobjects(1, num_figs);
+            f_output(run_figs) = f_output_run;
+
+            t_output = cell(1,num_tabs);
+            t_output(run_tabs) = t_output_run;
+
+            testCase.fig_success = f_success;
+            testCase.tab_success = t_success;
+            testCase.fig_output  = f_output;
+            testCase.tab_output  = t_output;
+        end
     end
 
     methods(TestClassTeardown)
@@ -119,33 +158,41 @@ classdef (SharedTestFixtures={ ...
                 testCase.assumeFalse( ismember(which_tabs, test.slow_tabs) );
             end
 
-            [success_criterion,fig_out,tab_out] = all_figures(which_figs,which_tabs,testCase.uuid.Value);
-
-            if isempty(success_criterion)
-                success_criterion = 1;
-            else
-                success_criterion = success_criterion{:};
-            end
-
             if which_figs ~= 0 % figure
+                success_criterion = testCase.fig_success{which_figs};
+                fig_out = testCase.fig_output(which_figs);
+
                 fig_name = ['Figure_' num2str(which_figs)];
                 pdf_name = ['../test-results/' fig_name];
                 
-                if ~isempty(fig_out.UserData)
-                    % pdf already exists in files, just copy to folder
-                    copyfile(fig_out.UserData, pdf_name)
-                else
-                    % save pdf from matlab figure output
-                    save_pdf(fig_out,pdf_name)
-                end
-                % in either case, use figure itself, not pdf, for printing the diagnostic
-                diagnostic = matlab.unittest.diagnostics.FigureDiagnostic(fig_out,'Prefix',[fig_name '_']);
+                if ~isa(fig_out,'matlab.graphics.GraphicsPlaceholder') % if figure exists (didn't error first)
+                    if ~isempty(fig_out.UserData)
+                        % pdf already exists in files, just copy to folder
+                        copyfile(fig_out.UserData, pdf_name)
+                    else
+                        % save pdf from matlab figure output
+                        save_pdf(fig_out,pdf_name)
+                    end
+                    % in either case, use figure itself, not pdf, for printing the diagnostic
+                    diagnostic = matlab.unittest.diagnostics.FigureDiagnostic(fig_out,'Prefix',[fig_name '_']);
+                end   
 
             else % table
+                success_criterion = testCase.tab_success{which_tabs};
+                tab_out = testCase.tab_output(which_tabs);
+
                 diagnostic = matlab.unittest.diagnostics.DisplayDiagnostic(tab_out{:});
             end
 
-            testCase.verifyGreaterThan(success_criterion, 0, diagnostic);
+            if isempty(success_criterion)
+                success_criterion = 1;
+            end
+
+            if isa(success_criterion,'MException')
+                rethrow(success_criterion)
+            else
+                testCase.verifyGreaterThan(success_criterion, 0, diagnostic);
+            end
 
         end
